feat(*): 增加BeforeQueryCallback

+++ sys_callback.go
@@ -9,6 +9,9 @@ func registerCallbacks() {
 	if gorm.DefaultCallback.Query().Get("kuu:query") == nil {
 		gorm.DefaultCallback.Query().Before("gorm:query").Register("kuu:query", QueryCallback)
 	}
+	if gorm.DefaultCallback.Query().Get("kuu:before_query") == nil {
+		gorm.DefaultCallback.Query().Before("gorm:query").Register("kuu:before_query", BeforeQueryCallback)
+	}
 	if gorm.DefaultCallback.Update().Get("kuu:update") == nil {
 		gorm.DefaultCallback.Update().Before("gorm:update").Register("kuu:update", UpdateCallback)
 	}
@@ -18,24 +21,36 @@ func registerCallbacks() {
 	if gorm.DefaultCallback.Create().Get("kuu:create") == nil {
 		gorm.DefaultCallback.Create().Before("gorm:create").Register("kuu:create", CreateCallback)
 	}
+	if gorm.DefaultCallback.Create().Get("kuu:create") == nil {
+		gorm.DefaultCallback.Create().Before("gorm:create").Register("kuu:create", CreateCallback)
+	}
+}
+
+// BeforeQueryCallback
+var BeforeQueryCallback = func(scope *gorm.Scope) {
+	if !scope.HasError() {
+		scope.CallMethod("BeforeFind")
+	}
 }
 
 // CreateCallback
 var CreateCallback = func(scope *gorm.Scope) {
-	if v, ok := GetValue(PrisDescKey); ok && v != nil {
-		desc := v.(*PrivilegesDesc)
-		if desc.IsValid() && !scope.HasError() {
-			if orgIDField, ok := scope.FieldByName("OrgID"); ok {
-				if orgIDField.IsBlank {
-					orgIDField.Set(desc.SignOrgID)
+	if !scope.HasError() {
+		if v, ok := GetValue(PrisDescKey); ok && v != nil {
+			desc := v.(*PrivilegesDesc)
+			if desc.IsValid() && !scope.HasError() {
+				if orgIDField, ok := scope.FieldByName("OrgID"); ok {
+					if orgIDField.IsBlank {
+						orgIDField.Set(desc.SignOrgID)
+					}
 				}
-			}
-			if createdByField, ok := scope.FieldByName("CreatedByID"); ok {
-				createdByField.Set(desc.UID)
-			}
+				if createdByField, ok := scope.FieldByName("CreatedByID"); ok {
+					createdByField.Set(desc.UID)
+				}
 
-			if updatedByField, ok := scope.FieldByName("UpdatedByID"); ok {
-				updatedByField.Set(desc.UID)
+				if updatedByField, ok := scope.FieldByName("UpdatedByID"); ok {
+					updatedByField.Set(desc.UID)
+				}
 			}
 		}
 	}
@@ -43,23 +58,25 @@ var CreateCallback = func(scope *gorm.Scope) {
 
 // DeleteCallback
 var DeleteCallback = func(scope *gorm.Scope) {
-	if v, ok := GetValue(PrisDescKey); ok && v != nil {
-		desc := v.(*PrivilegesDesc)
-		if desc.IsValid() && !scope.HasError() {
-			var extraOption string
-			if str, ok := scope.Get("gorm:delete_option"); ok {
-				extraOption = fmt.Sprint(str)
-			}
-			deletedByField, hasDeletedByField := scope.FieldByName("DeletedByID")
-			if !scope.Search.Unscoped && hasDeletedByField {
-				scope.Raw(fmt.Sprintf(
-					"UPDATE %v SET %v=%v%v%v",
-					scope.QuotedTableName(),
-					scope.Quote(deletedByField.DBName),
-					scope.AddToVars(desc.UID),
-					AddExtraSpaceIfExist(scope.CombinedConditionSql()),
-					AddExtraSpaceIfExist(extraOption),
-				)).Exec()
+	if !scope.HasError() {
+		if v, ok := GetValue(PrisDescKey); ok && v != nil {
+			desc := v.(*PrivilegesDesc)
+			if desc.IsValid() && !scope.HasError() {
+				var extraOption string
+				if str, ok := scope.Get("gorm:delete_option"); ok {
+					extraOption = fmt.Sprint(str)
+				}
+				deletedByField, hasDeletedByField := scope.FieldByName("DeletedByID")
+				if !scope.Search.Unscoped && hasDeletedByField {
+					scope.Raw(fmt.Sprintf(
+						"UPDATE %v SET %v=%v%v%v",
+						scope.QuotedTableName(),
+						scope.Quote(deletedByField.DBName),
+						scope.AddToVars(desc.UID),
+						AddExtraSpaceIfExist(scope.CombinedConditionSql()),
+						AddExtraSpaceIfExist(extraOption),
+					)).Exec()
+				}
 			}
 		}
 	}
@@ -75,33 +92,37 @@ func AddExtraSpaceIfExist(str string) string {
 
 // UpdateCallback
 var UpdateCallback = func(scope *gorm.Scope) {
-	if v, ok := GetValue(PrisDescKey); ok && v != nil {
-		desc := v.(*PrivilegesDesc)
-		if desc.IsValid() {
-			scope.SetColumn("UpdatedByID", desc.UID)
+	if !scope.HasError() {
+		if v, ok := GetValue(PrisDescKey); ok && v != nil {
+			desc := v.(*PrivilegesDesc)
+			if desc.IsValid() {
+				scope.SetColumn("UpdatedByID", desc.UID)
+			}
 		}
 	}
 }
 
 // QueryCallback
 var QueryCallback = func(scope *gorm.Scope) {
-	rawDesc, _ := GetValue(PrisDescKey)
-	rawValues, _ := GetValue(ValuesKey)
+	if !scope.HasError() {
+		rawDesc, _ := GetValue(PrisDescKey)
+		rawValues, _ := GetValue(ValuesKey)
 
-	if !IsBlank(rawDesc) {
-		if !IsBlank(rawValues) {
-			values := make(Values)
-			values = *(rawValues.(*Values))
-			if _, ok := values[IgnoreAuthKey]; ok {
-				return
+		if !IsBlank(rawDesc) {
+			if !IsBlank(rawValues) {
+				values := make(Values)
+				values = *(rawValues.(*Values))
+				if _, ok := values[IgnoreAuthKey]; ok {
+					return
+				}
 			}
-		}
-		desc := rawDesc.(*PrivilegesDesc)
-		if desc.NotRootUser() {
-			_, hasOrgIDField := scope.FieldByName("OrgID")
-			_, hasCreatedByIDField := scope.FieldByName("CreatedByID")
-			if hasOrgIDField && hasCreatedByIDField {
-				scope.Search.Where("(org_id IS NULL) OR (org_id = 0) OR (org_id in (?)) OR (created_by_id = ?)", desc.ReadableOrgIDs, desc.UID)
+			desc := rawDesc.(*PrivilegesDesc)
+			if desc.NotRootUser() {
+				_, hasOrgIDField := scope.FieldByName("OrgID")
+				_, hasCreatedByIDField := scope.FieldByName("CreatedByID")
+				if hasOrgIDField && hasCreatedByIDField {
+					scope.Search.Where("(org_id IS NULL) OR (org_id = 0) OR (org_id in (?)) OR (created_by_id = ?)", desc.ReadableOrgIDs, desc.UID)
+				}
 			}
 		}
 	}
